fix: restructure and requested changes
