+++ backend/src/services/secret-sync/render/render-sync-fns.ts
@@ -112,7 +112,7 @@ const redeployService = async (secretSync: TRenderSyncWithCredentials) => {
       {
         headers: {
           Authorization: `Bearer ${apiKey}`,
-          "Accept-Encoding": "application/json"
+          Accept: "application/json"
         }
       }
     )
