+++ frontend/src/hooks/api/secretSyncs/types/render-sync.ts
@@ -1,6 +1,6 @@
 import { AppConnection } from "@app/hooks/api/appConnections/enums";
 import { SecretSync } from "@app/hooks/api/secretSyncs";
-import { TRootSecretSync } from "@app/hooks/api/secretSyncs/types/root-sync";
+import { RootSyncOptions, TRootSecretSync } from "@app/hooks/api/secretSyncs/types/root-sync";
 
 export type TRenderSync = TRootSecretSync & {
   destination: SecretSync.Render;
@@ -17,7 +17,7 @@ export type TRenderSync = TRootSecretSync & {
     id: string;
   };
 
-  syncOptions: {
+  syncOptions: RootSyncOptions & {
     autoRedeployServices?: boolean;
   };
 };
