+++ frontend/src/pages/secret-manager/SecretSyncDetailsByIDPage/components/SecretSyncOptionsSection/RenderSyncOptionsSection.tsx
@@ -1,6 +1,6 @@
 import { GenericFieldLabel } from "@app/components/secret-syncs";
 import { Badge } from "@app/components/v2";
-import { TRenderSync } from "@app/hooks/api/secretSyncs/render-sync";
+import { TRenderSync } from "@app/hooks/api/secretSyncs/types/render-sync";
 
 type Props = {
   secretSync: TRenderSync;
@@ -12,16 +12,10 @@ export const RenderSyncOptionsSection = ({ secretSync }: Props) => {
   } = secretSync;
 
   return (
-    <div>
-      {autoRedeployServices ? (
-        <GenericFieldLabel label="Auto Redeploy Services">
-          <Badge variant="success">Enabled</Badge>
-        </GenericFieldLabel>
-      ) : (
-        <GenericFieldLabel label="Auto Redeploy Services">
-          <Badge variant="danger">Disabled</Badge>
-        </GenericFieldLabel>
-      )}
-    </div>
+    <GenericFieldLabel label="Auto Redeploy Services">
+      <Badge variant={autoRedeployServices ? "success" : "danger"}>
+        {autoRedeployServices ? "Enabled" : "Disabled"}
+      </Badge>
+    </GenericFieldLabel>
   );
 };
